+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### 修改
+
+- `KeyRangeMap`添加`into_iter`, `retain`, `remove`方法
+
 ## [0.5.0] 2021.12-30
 
 ### 添加